transactie naar eigen bestand, verbeteringen voor toevoegen synoniemen
